encapsulated timing functions on ConsensusService
