--- neo/Consensus/ConsensusTimer.cs
@@ -1,8 +1,0 @@
-namespace Neo.Consensus
-{
-    internal class ConsensusTimer
-    {
-        public uint Height;
-        public byte ViewNumber;
-    }
-}
